refactor: turn landingFooter.jsx into .tsx
